==> notebooks/02a_MTAT_SPLIT.ipynb <==
import numpy as np
import pandas as pd
import random
import torch
import os

from sklearn.model_selection import train_test_split

def set_seed(seed: int = 42) -> None:
    np.random.seed(seed)
    random.seed(seed)
    torch.manual_seed(seed)
    torch.cuda.manual_seed(seed)
    # Set a fixed value for the hash seed
    os.environ["PYTHONHASHSEED"] = str(seed)
    print(f"Random seed set as {seed}")

set_seed(123456)

split_path = "../split/mtat-20/"
os.makedirs(split_path, exist_ok=True)

broken_tracks = [
    "6/norine_braun-now_and_zen-08-gently-117-146.mp3",
    "8/jacob_heringman-josquin_des_prez_lute_settings-19-gintzler__pater_noster-204-233.mp3",
    "9/american_baroque-dances_and_suites_of_rameau_and_couperin-25-le_petit_rien_xiveme_ordre_couperin-88-117.mp3",
    "c/domased-new_memories-07-first_sun_rays-233-262.mp3",
    "d/tim_rayborn-chordae-04-sempr_alegria-784-813.mp3"
]

adnotations = pd.read_csv('../data/mtat/annotations_final.csv', sep='\t', index_col="clip_id")
adnotations = adnotations[~adnotations["mp3_path"].isin(broken_tracks)]
adnotations

clip_info = pd.read_csv('../data/mtat/clip_info_final.csv', sep='\t')
clip_info

# n tags with the largest quantity in dataset

n = 20

tags = adnotations.filter(regex='[^mp3_path]').sum().nlargest(20)
tags

tags = tags.keys().to_numpy('str')
np.save(open(os.path.join(split_path, 'tags.npy'), 'wb'), tags)
tags

# Binary tags

binary = adnotations[tags]
binary

indexes_to_keep = binary[binary.sum(axis=1) != 0].index.values
indexes_to_keep

binary = binary[binary.index.isin(indexes_to_keep)]
binary.insert(0, "clip_id", binary.index)
binary

binary_np = binary.to_numpy()
np.save(os.path.join(split_path, 'binary.npy'), binary_np)
binary_np

adnotations = adnotations[adnotations.index.isin(indexes_to_keep)]
adnotations = adnotations['mp3_path']
adnotations

# Spliting dataset to train, valid and test datasets

dataset = pd.DataFrame(adnotations, columns=['idx', 'mp3_path'])
dataset["idx"] = dataset.index
dataset

dataset.to_numpy()

train_ratio = 0.75
valid_ratio = 0.15
test_ratio = 0.1

train, test = train_test_split(dataset, test_size=1 - train_ratio)
valid, test = train_test_split(test, test_size=test_ratio/(test_ratio + valid_ratio))

dataset_train = train.to_numpy()
np.save(open(os.path.join(split_path, 'train.npy'), 'wb'), dataset_train)
dataset_train

dataset_valid = valid.to_numpy()
np.save(open(os.path.join(split_path, 'valid.npy'), 'wb'), dataset_valid)
dataset_valid

dataset_test = test.to_numpy()
np.save(open(os.path.join(split_path, 'test.npy'), 'wb'), dataset_test)
dataset_test
