==> notebooks/06_Train_MusicNN_20-class.ipynb <==
%pwd

import os
import random
import numpy as np
import torch

from external.musicnn import Musicnn
from utils.config import Config
from components.trainer import Trainer
from components.tester import Tester

config = Config(model=Musicnn(n_class=20),
                # preprocessor=PreProcessor(input_path="../data/mtat/mp3", output_path="../data/mtat/npy", suffix="npy", sr=16000),
                # preprocessor=OpenL3PreProcessor(input_path="../data/mtat/mp3", output_path="../data/mtat/emb", suffix="npy", sr=16000),
                n_epochs=40,
                batch_size=16,
                lr=1e-4,
                dataset_split_path="../split",
                model_filename_path="../models",
                dataset_name="mtat-20",
                data_path='../data',
                log_step=100,
                sr=16000,
                input_length=3 * 16000)

def set_seed(seed: int = 42) -> None:
    np.random.seed(seed)
    random.seed(seed)
    torch.manual_seed(seed)
    torch.cuda.manual_seed(seed)
    # Set a fixed value for the hash seed
    os.environ["PYTHONHASHSEED"] = str(seed)
    print(f"Random seed set as {seed}")

set_seed(123456)

trainer = Trainer(config)
trainer.train()

tester = Tester(config, trainer.model_file_name)
tester.test()
