==> notebooks/07_OpenL3_KNeighborsClassifier.ipynb <==
import os
import random
import numpy as np
import torch
import pathlib
from tqdm import tqdm

from utils.common import load_file_lists

def create_features(features_path, split_path):
    try:
        X_train = np.load(os.path.join(features_path, "X_train.npy"), allow_pickle=True)
        Y_train = np.load(os.path.join(features_path, "Y_train.npy"), allow_pickle=True)
        X_test = np.load(os.path.join(features_path, "X_test.npy"), allow_pickle=True)
        Y_test = np.load(os.path.join(features_path, "Y_test.npy"), allow_pickle=True)
    except OSError:
        binary = {row[0]: row[1:] for row in np.load(os.path.join(split_path, "binary.npy"), allow_pickle=True)}

        X_train = []
        Y_train = []
        test_data = load_file_lists([os.path.join(split_path, "train.npy"), os.path.join(split_path, "valid.npy")])
        for idx, filename in tqdm(test_data):
            filename = os.path.join("../data/mtat/emb", str(pathlib.Path(filename).with_suffix(".npy")))
            file_data = np.load(filename, allow_pickle=True).flatten()
            X_train.append(file_data)
            Y_train.append(binary[int(idx)])
        X_train = np.array(X_train)
        Y_train = np.array(Y_train)
        np.save(os.path.join(features_path, "X_train.npy"), Y_train)
        np.save(os.path.join(features_path, "Y_train.npy"), Y_train)

        X_test = []
        Y_test = []
        test_data = load_file_lists([os.path.join(split_path, "test.npy")])
        for idx, filename in tqdm(test_data):
            filename = os.path.join("../data/mtat/emb", str(pathlib.Path(filename).with_suffix(".npy")))
            file_data = np.load(filename, allow_pickle=True).flatten()
            X_test.append(file_data)
            Y_test.append(binary[int(idx)])
        X_test = np.array(X_test)
        Y_test = np.array(Y_test)
        np.save(os.path.join(features_path, "X_test.npy"), X_test)
        np.save(os.path.join(features_path, "Y_test.npy"), Y_test)
    return X_train, X_test, Y_train, Y_test

def set_seed(seed: int = 42) -> None:
    np.random.seed(seed)
    random.seed(seed)
    torch.manual_seed(seed)
    torch.cuda.manual_seed(seed)
    os.environ["PYTHONHASHSEED"] = str(seed)
    print(f"Random seed set as {seed}")

set_seed(123456)

dataset_name = "mtat-10"
split_path = "../split"
features_path = "../data/mtat/features/"

split_path = os.path.join(split_path, dataset_name)
features_path = os.path.join(features_path, dataset_name)
os.makedirs(split_path, exist_ok=True)
os.makedirs(features_path, exist_ok=True)

X_train, X_test, Y_train, Y_test = create_features(features_path, split_path)

X_train

print(X_train.shape)
print(X_test.shape)
print(Y_train.shape)
print(Y_test.shape)

import pickle
from sklearn.exceptions import UndefinedMetricWarning
from sklearn.metrics import classification_report
from sklearn.neighbors import KNeighborsClassifier

import warnings
warnings.simplefilter(action='ignore', category=UndefinedMetricWarning)

model = KNeighborsClassifier()
model.fit(X_train, Y_train)
y_pred = model.predict(X_test)
print(classification_report(y_pred, Y_test))

model_filename_path = os.path.join("../models", model.__class__.__name__)
os.makedirs(model_filename_path, exist_ok=True)
model_filename = os.path.join(model_filename_path, f"{dataset_name}.bin")
pickle.dump(model, open(model_filename, 'wb+'))

loaded_model = pickle.load(open("../models/KNeighborsClassifier/mtat-10.bin", 'rb'))
y_pred = loaded_model.predict(X_test)
print(classification_report(y_pred, Y_test))

import pandas as pd
from src.components.preprocessor import OpenL3PreProcessor
from src.utils.config import Config


class SklearnPredictor:
    def __init__(self, config: Config, model_filename: str = None, cuda: bool = None):
        if config is None:
            config = Config()

        tags_path = os.path.join(config.dataset_split_path, config.dataset_name, "tags.npy")
        self.tags = np.load(tags_path, allow_pickle=True)

        self.model_filename_path = config.model_filename_path
        self.model_filename = model_filename

        # cuda
        self.is_cuda = torch.cuda.is_available() if cuda is None else cuda

        self.data_path = config.data_path
        self.input_length = config.input_length

        self.preprocessor = OpenL3PreProcessor()

    def predict_tags(self, mp3_file, model=None) -> pd.DataFrame:
        if model is None:
            model = pickle.load(open(os.path.join(self.model_filename_path, self.model_filename), "rb"))

        out = self.predict_file(mp3_file, model).T
        df = pd.DataFrame(out, index=self.tags)
        df = df.reindex(df.mean(axis=1).sort_values(ascending=False).index)
        return df

    def predict_file(self, mp3_file, model=None):
        if model is None:
            model = pickle.load(open(os.path.join(self.model_filename_path, self.model_filename), "rb"))

        return self.predict_data(self.preprocessor.process(mp3_file), model)

    def predict_data(self, x, model=None):
        if model is None:
            model = pickle.load(open(os.path.join(self.model_filename_path, self.model_filename), "rb"))

        return model.predict(x.reshape(1, -1))

file = "../data/mtat/mp3/6/barry_phillips-trad-03-bisonpolska-88-117.mp3"
predictor = SklearnPredictor(Config(dataset_split_path="../split", model_filename_path="../models", dataset_name="mtat-10"),
                             model_filename="KNeighborsClassifier/mtat-10/model.bin")

predictor.predict_file(file)
