==> notebooks/08_Data_Expore.ipynb <==
# Imports

import os
import pandas as pd

# Settings

file = "../data/mtat/mp3/8/beth_quist-lucidity-01-crazy-204-233.mp3"
sr = 16000
broken_tracks = [
    "6/norine_braun-now_and_zen-08-gently-117-146.mp3",
    "8/jacob_heringman-josquin_des_prez_lute_settings-19-gintzler__pater_noster-204-233.mp3",
    "9/american_baroque-dances_and_suites_of_rameau_and_couperin-25-le_petit_rien_xiveme_ordre_couperin-88-117.mp3",
    "c/domased-new_memories-07-first_sun_rays-233-262.mp3",
    "d/tim_rayborn-chordae-04-sempr_alegria-784-813.mp3"
]

if not os.path.exists("images"):
    os.mkdir("images")

adnotations = pd.read_csv('../data/mtat/annotations_final.csv', sep='\t', index_col="clip_id")
adnotations = adnotations[~adnotations["mp3_path"].isin(broken_tracks)]
adnotations = adnotations.filter(regex='[^mp3_path]')
adnotations

all_tags = adnotations.sum()
all_tags

## Analiza dla 20 najbardziej licznych tagów

tags = adnotations.sum().nlargest(20)
tags

adnotations_n = adnotations[tags.keys().to_numpy('str')]
indexes_to_keep = adnotations_n[adnotations_n.sum(axis=1) != 0].index.values
adnotations_n = adnotations_n[adnotations_n.index.isin(indexes_to_keep)]
adnotations_n

### Liczba tagów

tags_count = tags.to_frame()
tags_count

import plotly.graph_objs as go

data = [go.Bar(y=tags_count.index,
               x=tags_count[0],
               text=tags_count[0],
               textposition="outside",
               orientation = 'h')]
layout = ({"title": {"text":"Liczba utworów otagowanych przez określoną etykietę (30 najbardziej licznych etykiet)", "x": 0.5},
           "yaxis": {"title":"Etykieta", "autorange": "reversed"},
           "xaxis": {"title":"Liczba utwórów"},
           "showlegend": False,
           "template": "plotly_white"})

fig = go.Figure(data = data, layout = layout)
fig.show()
fig.write_image("images/explore-fig1.png", width=1000, height=750)

### Liczba unikalnych tagów (tylko 1 przy utworze)

unique_tags = adnotations_n[adnotations_n.sum(axis=1) == 1]
unique_tags = unique_tags.sum().sort_values(ascending=False).to_frame()
unique_tags

import plotly.graph_objs as go

data = [go.Bar(y=unique_tags.index,
               x=unique_tags[0],
               text=unique_tags[0],
               textposition="outside",
               orientation = 'h')]
layout = ({"title": {"text":"Liczba utworów otagowanych przez tylko jedną etykietę (30 najbardziej licznych etykiet)", "x": 0.5},
           "yaxis": {"title":"Etykieta", "autorange": "reversed"},
           "xaxis": {"title":"Liczba utwórów"},
           "showlegend": False,
           "template": "plotly_white"})

fig = go.Figure(data = data, layout = layout)
fig.show()
fig.write_image("images/explore-fig2.png", width=1000, height=750)

tags_unique_ratio = tags_count.rename(columns={0: "all"}).join(unique_tags.rename(columns={0: "unique"}))
tags_unique_ratio["ratio"] = tags_unique_ratio["unique"] / tags_unique_ratio["all"]
tags_unique_ratio = tags_unique_ratio.sort_values(by=['ratio'], ascending=False)
tags_unique_ratio

import plotly.graph_objs as go

data = [go.Bar(y=tags_unique_ratio.index,
               x=tags_unique_ratio.ratio,
               text=round(tags_unique_ratio.ratio, 4),
               textposition="outside",
               orientation = 'h')]
layout = ({"title": {"text":"Stosunek liczby utworów z unikalnym tagiem do wszystkich utworów etykietowanych tym tagiem", "x": 0.5},
           "yaxis": {"title":"Etykieta", "autorange": "reversed"},
           "xaxis": {"title":"Utwory z unikalnym tagiem / wszystkie utwory z tagiem"},
           "showlegend": False,
           "template": "plotly_white"})

fig = go.Figure(data = data, layout = layout)
fig.show()
fig.write_image("images/explore-fig4.png", width=1100, height=750)

### Liczba etykiet przy utworze

import numpy as np
tags_hist = adnotations_n.sum(axis=1).sort_values(ascending=False).to_frame()
tags_hist

print(f"Min: {np.round(np.min(tags_hist.values), 2)}")
print(f"Mean: {np.round(np.mean(tags_hist.values), 2)}")
print(f"Median: {np.round(np.median(tags_hist.values), 2)}")
print(f"Std: {np.round(np.std(tags_hist.values), 2)}")
print(f"Max: {np.round(np.max(tags_hist.values), 2)}")

import plotly.graph_objs as go

counts, bins = np.histogram(tags_hist[0], bins=range(1, 12))
data = [go.Bar(x=bins, y=counts, text=counts, textposition="outside")]
layout = ({"title": {"text":"Liczba tagów przypisanych do utworu (30 najbardziej licznych etykiet)", "x": 0.5},
           "yaxis": {"title":"Liczba utworów"},
           "xaxis": {"title":"Liczba tagów", "tickmode": 'linear'},
           "showlegend": False,
           "template": "plotly_white"})

fig = go.Figure(data = data, layout = layout)
fig.update_traces(textposition='outside')
fig.show()
fig.write_image("images/explore-fig3.png", width=750, height=500)

# MelSpectrogram

from torch import Tensor, flatten
from utils.common import convert_mp3_to_npy
from torchaudio.transforms import AmplitudeToDB, MelSpectrogram

to_db = AmplitudeToDB()
spec = MelSpectrogram(sample_rate=sr,
                             n_fft=512,
                             f_min=0.0,
                             f_max=8000.0,
                             n_mels=50)
data = convert_mp3_to_npy(file, sr)
data = Tensor(data.reshape(1, -1))
data = spec(data)
print(data.shape)
print(flatten(data).shape)
data

# OpenL3

from components.preprocessor import OpenL3PreProcessor

preprocessor = OpenL3PreProcessor()
data = preprocessor.process(file)
print(data.shape)
print(data.flatten().shape)
data
