==> notebooks/09_Results.ipynb <==
from sklearn.neighbors import KNeighborsClassifier

from utils.config import Config
from components.predictor import SklearnPredictor

sklearn_config = Config(model=KNeighborsClassifier(),
                model_filename_path = "../models",
                data_path='../data',
                dataset_split_path= "../split",
                dataset_name="mtat-20")
sklearn_predictor = SklearnPredictor(sklearn_config, model_filename="model.bin")

import numpy as np

test_files = np.load("../split/mtat-20/test.npy", allow_pickle=True)
tags = np.load("../split/mtat-20/tags.npy", allow_pickle=True)
binary = {row[0]: row[1:] for row in np.load("../split/mtat-20/binary.npy", allow_pickle=True)}

from components.preprocessor import OpenL3PreProcessor

preprocessor = OpenL3PreProcessor(input_path="../data/mtat/mp3", output_path="../data/mtat/emb", suffix="npy", sr=16000)
preprocessor.run(test_files[:, 1])

import pathlib
import os
from tqdm import tqdm

X = []
Y_true = []
for idx, file in tqdm(test_files):
    file = os.path.join("../data/mtat/emb/", pathlib.Path(file).with_suffix(f".npy"))
    X.append(np.load(file, allow_pickle=True))
    Y_true.append(binary[idx])
Y_true = np.array(Y_true)

Y_pred = []
for x in tqdm(X[:50]):
    Y_pred.append(sklearn_predictor.predict_data_prob(x))
Y_pred = np.array(Y_pred)

Y_pred = Y_pred.reshape(Y_pred.shape[0], -1)

Y_pred.shape

Y_true = Y_true[:50]

from sklearn.metrics import classification_report
from utils.common import Statistics

stats = Statistics(0, Y_pred, Y_true)
est_bin_array = np.where(stats.est_array >= 0.5, 1, 0)
print(classification_report(stats.gt_array, est_bin_array, target_names=list(sklearn_predictor.tags), zero_division=1))
