==> notebooks/11-Wplyw-Parametru-K.ipynb <==
import os
import random
import numpy as np
import torch
import pathlib
from tqdm import tqdm

from utils.common import load_file_lists

def create_features(split_path):
    binary = {row[0]: row[1:] for row in np.load(os.path.join(split_path, "binary.npy"), allow_pickle=True)}
    X_train = []
    Y_train = []
    test_data = load_file_lists([os.path.join(split_path, "train.npy"), os.path.join(split_path, "valid.npy")])
    for idx, filename in tqdm(test_data):
        filename = os.path.join("../data/mtat/emb", str(pathlib.Path(filename).with_suffix(".npy")))
        file_data = np.load(filename, allow_pickle=True).flatten()
        X_train.append(file_data)
        Y_train.append(binary[int(idx)])
    X_train = np.array(X_train)
    Y_train = np.array(Y_train)
    return X_train, Y_train

def set_seed(seed: int = 42) -> None:
    np.random.seed(seed)
    random.seed(seed)
    torch.manual_seed(seed)
    torch.cuda.manual_seed(seed)
    os.environ["PYTHONHASHSEED"] = str(seed)
    print(f"Random seed set as {seed}")

set_seed(123456)

dataset_name = "mtat-20"
split_path = "../split"
features_path = "../data/mtat/features/"

split_path = os.path.join(split_path, dataset_name)
os.makedirs(split_path, exist_ok=True)
X_train, Y_train = create_features(split_path)

X_train

Y_train

from components.tester import SklearnTester
from utils.config import Config
from sklearn.exceptions import UndefinedMetricWarning
from sklearn.neighbors import KNeighborsClassifier
import warnings
warnings.simplefilter(action='ignore', category=UndefinedMetricWarning)

config = Config(model=KNeighborsClassifier(), model_filename_path="../models", data_path='../data',
                dataset_split_path="../split", dataset_name=f"mtat-20")
tester = SklearnTester(config)

results_20 = {}
for i in range(1, 16):
    print(f"K = {i}")
    model = KNeighborsClassifier(n_neighbors=i)
    model.fit(X_train, Y_train)
    stats = tester.test(model)
    results_20[i] = [stats.f1_score, stats.roc_auc, stats.pr_auc]

dataset_name = "mtat-10"
split_path = "../split"
features_path = "../data/mtat/features/"

split_path = os.path.join(split_path, dataset_name)
os.makedirs(split_path, exist_ok=True)
X_train, Y_train = create_features(split_path)

X_train

Y_train

from components.tester import SklearnTester
from utils.config import Config
from sklearn.exceptions import UndefinedMetricWarning
from sklearn.neighbors import KNeighborsClassifier
import warnings
warnings.simplefilter(action='ignore', category=UndefinedMetricWarning)

config = Config(model=KNeighborsClassifier(), model_filename_path="../models", data_path='../data',
                dataset_split_path="../split", dataset_name=f"mtat-10")
tester = SklearnTester(config)

results_10 = {}
for i in range(1, 16):
    print(f"K = {i}")
    model = KNeighborsClassifier(n_neighbors=i)
    model.fit(X_train, Y_train)
    stats = tester.test(model)
    results_10[i] = [stats.f1_score, stats.roc_auc, stats.pr_auc]

import pandas as pd
df_10 = pd.DataFrame.from_dict(results_10, orient="index", columns=["f1_score", "auc_roc", "auc_pr"])
df_10

df_20 = pd.DataFrame.from_dict(results_20, orient="index", columns=["f1_score", "auc_roc", "auc_pr"])
df_20

from plotly.subplots import make_subplots
import plotly.graph_objects as go

fig = make_subplots(rows=1, cols=3, subplot_titles=['F1 Score', 'Area under ROC curve', 'Area under PR curve'])
fig.add_trace(go.Scatter(x=df_10.index, y=df_10["f1_score"], name='mtat-10'), col=1, row=1)
fig.add_trace(go.Scatter(x=df_20.index, y=df_20["f1_score"], name='mtat-20'), col=1, row=1)
fig.add_trace(go.Scatter(x=df_10.index, y=df_10["auc_roc"], name='mtat-10'), col=2, row=1)
fig.add_trace(go.Scatter(x=df_20.index, y=df_20["auc_roc"], name='mtat-20'), col=2, row=1)
fig.add_trace(go.Scatter(x=df_10.index, y=df_10["auc_pr"], name='mtat-10'), col=3, row=1)
fig.add_trace(go.Scatter(x=df_20.index, y=df_20["auc_pr"], name='mtat-20'), col=3, row=1)

fig.update_layout(({"title": {"text": "Strojenie parametru K", "x": 0.5},
                    "showlegend": True,
                    "template": "plotly_white"}))
fig.show()
fig.write_image("images/fig9.png", width=1000, height=500)

import pickle
from components.tester import SklearnTester
from utils.config import Config
from sklearn.exceptions import UndefinedMetricWarning
from sklearn.neighbors import KNeighborsClassifier
import warnings
warnings.simplefilter(action='ignore', category=UndefinedMetricWarning)

for n in [10, 20]:
    config = Config(model=KNeighborsClassifier(), model_filename_path="../models", data_path='../data',
                    dataset_split_path="../split", dataset_name=f"mtat-{n}")
    tester = SklearnTester(config)

    dataset_name = f"mtat-{n}"
    split_path = "../split"
    features_path = "../data/mtat/features/"

    split_path = os.path.join(split_path, dataset_name)
    os.makedirs(split_path, exist_ok=True)
    X_train, Y_train = create_features(split_path)

    model = KNeighborsClassifier(n_neighbors=4)
    model.fit(X_train, Y_train)
    model_filename = os.path.join("../models", model.__class__.__name__, config.dataset_name, "model.bin")
    pickle.dump(model, open(model_filename, 'wb+'))
    stats = tester.test(model)
